// src/app/models/api.models.ts
export interface CartItem {
  productId: string;
  productName: string;
  price: number;
  quantity: number;
  image: string;
  type: 'PURCHASE' | 'RENTAL';
  rentalDays?: number;
}